% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_game.R
\name{is_mxo_game}
\alias{is_mxo_game}
\title{Test whether an object is an \code{mxo_game}}
\usage{
is_mxo_game(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is an \code{mxo_game}
}
\examples{
is_mxo_game(mxo_new_game())
is_mxo_game(list())
}
