% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_policy.R
\name{is_mxo_policy}
\alias{is_mxo_policy}
\title{Test whether an object is a multixoR policy}
\usage{
is_mxo_policy(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a multixoR policy
}
