% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multixoR-package.R
\docType{package}
\name{multixoR-package}
\alias{multixoR}
\alias{multixoR-package}
\title{multixoR: Five-Dimensional Multiverse Tic-Tac-Toe}
\description{
A generic n^d lattice game engine that implements a multiverse variant of
tic-tac-toe across three spatial axes, time, and timeline. See the package
vignettes and the game specification for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/multixoR}
  \item \url{https://r-heller.github.io/multixoR/}
  \item Report bugs at \url{https://github.com/r-heller/multixoR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
