% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_print.R
\name{mxo_board}
\alias{mxo_board}
\title{Access a single board from the multiverse}
\usage{
mxo_board(game, L, t)
}
\arguments{
\item{game}{An \code{mxo_game} object.}

\item{L}{Integer scalar, timeline label.}

\item{t}{Integer scalar, time index.}
}
\value{
An integer vector of length \code{n^d_spatial} with values in
\verb{\{0L, 1L, 2L\}} (0 empty, 1 X, 2 O).
}
\description{
Access a single board from the multiverse
}
\examples{
mxo_board(mxo_new_game(), 0L, 0L)
}
