% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_calibration.R
\name{mxo_calibrator_predict}
\alias{mxo_calibrator_predict}
\title{Apply a calibrator to a numeric score}
\usage{
mxo_calibrator_predict(calibrator, score)
}
\arguments{
\item{calibrator}{An \code{mxo_calibrator}.}

\item{score}{Numeric vector of heuristic scores.}
}
\value{
Numeric vector of probabilities in \verb{[0, 1]}.
}
\description{
Apply a calibrator to a numeric score
}
