% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_win.R
\name{mxo_is_terminal}
\alias{mxo_is_terminal}
\title{Test whether the game has ended}
\usage{
mxo_is_terminal(game)
}
\arguments{
\item{game}{An \code{mxo_game} object.}
}
\value{
Logical scalar.
}
\description{
Test whether the game has ended
}
\examples{
mxo_is_terminal(mxo_new_game())
}
