% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_plot_eval.R
\name{mxo_plot_eval}
\alias{mxo_plot_eval}
\title{Plot the heuristic-evaluation curve along a game}
\usage{
mxo_plot_eval(record, ...)
}
\arguments{
\item{record}{An \code{mxo_game_record} from \code{\link[=mxo_self_play]{mxo_self_play()}}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the heuristic-evaluation curve along a game
}
