% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_plot_eval.R
\name{mxo_plot_opening}
\alias{mxo_plot_opening}
\title{Plot the openings heatmap}
\usage{
mxo_plot_opening(opening_table, n = 4L, d_spatial = 3L, ...)
}
\arguments{
\item{opening_table}{The tibble produced by \code{\link[=mxo_opening_table]{mxo_opening_table()}}.}

\item{n, d_spatial}{Geometry parameters used to derive cell coordinates.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Consumes the output of \code{\link[=mxo_opening_table]{mxo_opening_table()}} and renders X's first-move
win-rate as a faceted-by-z heatmap.
}
