% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_policy.R
\name{mxo_policy_move}
\alias{mxo_policy_move}
\title{Apply a policy to choose a move}
\usage{
mxo_policy_move(policy, game)
}
\arguments{
\item{policy}{An \code{mxo_policy} object.}

\item{game}{An \code{mxo_game} object.}
}
\value{
A one-row legal-moves tibble (or zero-row at terminal states).
}
\description{
Apply a policy to choose a move
}
