% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_win.R
\name{mxo_status}
\alias{mxo_status}
\title{Game status}
\usage{
mxo_status(game)
}
\arguments{
\item{game}{An \code{mxo_game} object.}
}
\value{
A list with components:
\itemize{
\item \code{status} (chr): one of \code{"in_progress"}, \code{"x_win"}, \code{"o_win"}, \code{"draw"}.
\item \code{winner} (int): \code{NA_integer_}, \code{1L} (X), or \code{2L} (O).
\item \code{win_line} (list or NULL): the winning cells if terminal-with-win.
}
}
\description{
Game status
}
\examples{
mxo_status(mxo_new_game())
}
