% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_print.R
\name{mxo_to_move}
\alias{mxo_to_move}
\title{Player to move}
\usage{
mxo_to_move(game)
}
\arguments{
\item{game}{An \code{mxo_game} object.}
}
\value{
Integer scalar, \code{1L} (X) or \code{2L} (O).
}
\description{
Player to move
}
\examples{
mxo_to_move(mxo_new_game())
}
