% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_calibration.R
\name{print.mxo_calibrator}
\alias{print.mxo_calibrator}
\title{Print a calibrator}
\usage{
\method{print}{mxo_calibrator}(x, ...)
}
\arguments{
\item{x}{An \code{mxo_calibrator}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a calibrator
}
