% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_print.R
\name{print.mxo_game}
\alias{print.mxo_game}
\alias{format.mxo_game}
\title{Print a multixoR game}
\usage{
\method{print}{mxo_game}(x, ...)

\method{format}{mxo_game}(x, ...)
}
\arguments{
\item{x}{An \code{mxo_game} object.}

\item{...}{Unused. Reserved for future arguments.}
}
\value{
Invisibly returns \code{x}.

For \code{format}, a character vector representation.
}
\description{
Print a multixoR game
}
\examples{
print(mxo_new_game())
}
