% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_mcts.R
\name{print.mxo_mcts_result}
\alias{print.mxo_mcts_result}
\title{Print an \code{mxo_mcts_result}}
\usage{
\method{print}{mxo_mcts_result}(x, top = 5L, ...)
}
\arguments{
\item{x}{An \code{mxo_mcts_result} object.}

\item{top}{Integer scalar, number of top-visited moves to display.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{mxo_mcts_result}
}
