% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_policy.R
\name{print.mxo_policy}
\alias{print.mxo_policy}
\title{Print a multixoR policy}
\usage{
\method{print}{mxo_policy}(x, ...)
}
\arguments{
\item{x}{An \code{mxo_policy}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a multixoR policy
}
