% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_simulate.R
\name{print.mxo_sim_result}
\alias{print.mxo_sim_result}
\title{Print an \code{mxo_sim_result}}
\usage{
\method{print}{mxo_sim_result}(x, ...)
}
\arguments{
\item{x}{An \code{mxo_sim_result}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{mxo_sim_result}
}
