% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxo_print.R
\name{summary.mxo_game}
\alias{summary.mxo_game}
\alias{print.mxo_game_summary}
\title{Summarise a multixoR game}
\usage{
\method{summary}{mxo_game}(object, ...)

\method{print}{mxo_game_summary}(x, ...)
}
\arguments{
\item{object}{An \code{mxo_game} object.}

\item{...}{Unused. Reserved for future arguments.}

\item{x}{An \code{mxo_game_summary} object.}
}
\value{
An object of class \code{mxo_game_summary} with counts per player and
branch information.

Invisibly returns \code{x}.
}
\description{
Summarise a multixoR game
}
\examples{
summary(mxo_new_game())
}
